Add small network support to zmap.
The command line will now accept an optional CIDR address allowing the
user to narrow the search to a specific netblock.  Any blacklist or
whitelist that is also specified modifies the probing space, but the user
isn't forced to use these files to scope the work for quick scans.

The general idea is that a table is implemented containing parameters for
various smaller cyclic group sizes based on the given subnet mask.  This
is much more efficient than cycling through the entire 0/0 space when
scanning a much smaller portion of the Internet.