Add clang-format to Travis
This adds a kernel style .clang-format configuration, and the build
commands `make format` and `make format-check`, which format and check
the formatting of ZMap code, respectively. The .clang-format file is for
clang-format 4.0. As such, during the TravisCI build, we only run make
format-check on OS X, since the clang-format included in the apt
repositories is out of date.