Add Github Action for compilation
This converts the TravisCI build to a Github Action. Similar to Travis,
we compile and run the script that checks that all options are defined
in the manfile. The action runs inside the new `zmap/builder` container,
which is defined in `\\container\builder.dockerfile`, and is (manually)
published to Dockerhub.

This action runs considerably faster than Travis (completes in less than
one minute).