Ensure all probe modules set *buf_len (#669)
* Ensure all probe modules set *buf_len

Probe modules with dynamic output already needed to set `*buf_len`
during calls to `make_packet`. Not all probe modules did this. Some
opted to update the value of `probe_module.packet_length`. This is not
safe in a multithreaded invocation.

This commit forces all probe modules to update `*buf_len`, and replaces
the packet_length parameter with a max_packet_length, which is used for
bandwidth rate calculation as an upper bound. It is unclear if the rate
calculation works currently, but at least it's getting consistent input
now.

* wip

* split templated from static udp

* fix templating

* remove debug logs