ipip: fix validation failure when setting src ip
When packets with source ip which is not the receiving address for
replies (e.g. behin NAT) validation data was computed with the actual
receiving address on receive. However, on send it is computed with the
sending address. This discrepancy made valid packets fail validation.

This commit fixes the discrepancy by trying every possible sender IP
when validating a packet.