Fix 32bit overflow in monitor.c (issue #715)
If number of probes is greater than 1, then any scan above 50%
will result in the multiplication in the given line to overflow.
This triggers a false minimum comparison, yielding wildly incorrect
scan time estimations.

The fix is to cast it to a uint64_t, as is done for the other clauses
above.