Stop defining the now defunct _SYSTYPE_BSD
This leftover bit was the result of a collision of two in-flight PRs.
The last use of _SYSTYPE_BSD was removed along with libdnet in #772,
while #771 concurrently added the -D_SYSTYPE_BSD to CMakeLists.txt.

_SYSTYPE_BSD is a rather archaic define.  libdnet might have used it for
compat with now historical systems/compilers, but I do not think it
served any purpose beyond that.  Certainly, modern FreeBSD, NetBSD and
macOS do not reference it.  It seems safe to remove.