Add --netmap-wait-ping to work around STP blocking ports temporarily
Going into and leaving netmap mode causes the link to go down and up as
part of a PHY reset.  If the interface is connected to a switch with STP
enabled, then depending on port configuration, the switch might be
muting the port for as many as 30 seconds while the port goes through
the listening and learning STP states.

To work around this in situations where the switch cannot be
reconfigured, `--netmap-wait-ping` allows to wait for end-to-end ICMP
connectivity with an IP address before starting the scan.