netmap: Reply to ARP requests from gateway for scan source IPs
In netmap mode, the OS network stack never gets to see incoming packets
unless we explicitly forward them to the host rings; hence the kernel
will not be responding to ARP requests.  To remove the need for static
ARP entries on the gateway, respond to ARP requests from the gateway for
any of the source IPs of the scan.