Probe-modules: IPIP (#684)
* Add ipip probe module

* IPIP: fix validate function

* ipip: fix validation failure when setting src ip

When packets with source ip which is not the receiving address for
replies (e.g. behin NAT) validation data was computed with the actual
receiving address on receive. However, on send it is computed with the
sending address. This discrepancy made valid packets fail validation.

This commit fixes the discrepancy by trying every possible sender IP
when validating a packet.

* Consistent indentation in probe_modules.c

* fixed up merge issues, added multi-port feature into ipip

* fixed up ipip probe function definitions and added ip id

---------

Co-authored-by: Yoav Alon <yoava333@gmail.com>
Co-authored-by: Zakir Durumeric <zakird@gmail.com>
Co-authored-by: phillip-stephens <phillip@cs.stanford.edu>