Add probe module prepare_packet callback to reduce copying
Introduce new prepare_packet callback for the initialization of send
buffers; contrary to the per-thread initialization callback where this
was done previously, prepare_packet can be called multiple times, once
for each send buffer.

Make use of this to prepare packets to send directly in the batch
buffers instead of copying them over.