Remove ip field from batch and improve alignment
The ip field was only still used in send-bsd and only on a failure path
for logging, which does not seem like a strong justification for keeping
it around, especially given that it can always be read from the packet
data itself.

While removing the ip field, align buf such that the IP header is going
to start at a 32 bit aligned address, which again results in 32 bit
header field access to be faster because of ideal alignment.